Create basic GameState semi-abstract class
Created a semi-abstract (as abstract as you can get in a language that
doesn't have an *abstract* keyword) class called GameState. As outlined
in code comments, the idea of a GameState is that we can easily swap
between, and isolate, different elements of code.

In order to test this idea, I moved most of the isolatable code into a
sample game state, found in src/sample_state.py. Everything seems to be
working as expected.