Revise start and stop functions
Previously, the start and stop functions in the semi-abstract class
pushed and popped the class method handlers. This lead to problems
for subclasses that overwrote the start and stop functions - which
they were supposed to. Basically, if you created a custom start
and/or stop function, you would have to:

        a) call the super's equivalent method, or

        b) duplicate the code from the super's method

To fix this, the offending code has been moved into semi-private
(Python private?) methods: _start and _stop, respectively.

Now,the code calls the private methods in the following sequence:

        1. _start
        2. start
        3. stop
        4. _stop

Whoevever creates a subclass will be responsible for cleaning up
whatever they do in the start & stop methods.