Add world generation foundation
Added a WorldMaker class, which will ultimately generate the world
according to a precise algorithm. Right now, it just generates some
stochastic noise.

The world currently comes in the form of a Python shared memory Array.
This will allow us to access and update the world using multiple
processes - much faster than doing things in a single process.

To access the Array in terms of x and y, we reshape it using numpy -
this has been added as a new requirement.