Refactor away temp code and fixes
Took a quick minute to refactor the code.

First, we removed sample_state.py, since that was just there for
demonstration.

Secondly, we moved the load_screen file into a different package.
Previously, it was in the city_state package. However, it was meant to
be generic - usable for whatever state-transition needs we have. It
made more sense, then, to have it in it's own package.

Finally, I removed the unecessary terrain.png file. This was
originally required because of how Pyglet's resource paths work: the
resource loader was only looking for ../assets/. This worked for
Python files in the src/ directory, but files in the lower packages
couldn't find the assets folder.

If you were in a package folder, it was looking in the folder above
it. This meant that, in city_state, we were looking for src/assets/,
instead of just assets/.

Adding an additional search path - "../../assets/". Of course, this
solution is kind of just a band-aid: we'll need to add a path for
each level of package depth. But that shouldn't be an issue - if we're
so many packages deep that it IS an issue then this project will be
in a very bad place.