Add view-based tile culling
Previously, we would just render the whole map once we were done
loading. This was good for testing, but would be impractical once we
began testing maps of any appreciable size. We also would need a way
to scroll around maps that are bigger than the screen size.

These have been implemented as the FlyCamera class, which handles the
scrolling for our map. Currently, the culling is implemented by
getting a list of visible coordinates, creating those tiles that
don't exist, and culling the tiles that do exist but are off screen.