Rework camera for efficiency and clarity
This commit fixes two issues with the "Camera" class:

	* Most of the variables were pixel-oriented. For example, both
	  the max and the min had to pixel values, calculated by
	  whatever was making the camera.

	* The only way to decide what tiles to render was to get all
	  the tiles currently on the screen.

Having everything be measured in terms of pixels was problematic,
because the atomic unit of the game world is the TILE, not the PIXEL.
This made things unecessarily clunky when creating the class.

Now, all but the view length/height are given in terms of tile size.
This rework has borked some older features but has generally made the
code more readable. The addition of property setters should make
working with the camera outside of the class easier, too.

Now, the issue with getting all the tiles on the screen constantly was
radical inefficiency. This would cause the program to chug when
scrolling around the world, or even sitting still.

The solution, then, was to create a camera that only provided lists
of the tiles that needed to be culled and created. This functionality
is provided by DiffCamera, which only reports new tiles and old tiles
from whenever the screen moves.

This camera code is not yet complete; it requires more notation
(especially on DiffCamera)and a few more features before we can move
on.