Make DiffCamera at least mildly functional
The DiffCamera is sorted of based on the idea of just gradually
modifying what's on screen, rather than wholesale changing it. That
works perfectly well so long as we can perfectly calculate our
changes.

The earlier version of DiffCamera was really inaccurate, and was
constantly missing tiles to make and tiles to cull.

The issue ended being that we were trying to check the change on the
various sides (left, right, top, bottom) all at once. Each one is
actually different and will typically report makes and culls at
separate times.

Each one of the different sides/borders now tracks it's own position
and reports whenever tiles need to be made or culled. To facilitate
this, the CameraBorder class has been added.

Some features of the TileCamera class had to be changed to be more
compatible with DiffCamera. First, the position attribute was removed
because we NEED to pass in the reset variable in DiffCamera. Also,
removed view_start argument, as it caused some issues when
initializing the CameraBorders in DiffCamera. It was redundant
anyway: we can just immediately set the position after creation.