Refactor camera-werks for tidyness
Took the time to refactor the Camera related classes.

Firstly, I wanted to add some more documentation and clean up the
existing documentation. Just in case we ever need to come back and
work more on the Camera class, I don't want it to have the (readily)
outward appearance of a madman's scribblings.

Secondly, I wanted to clean up the debug messages that I was using for
testing the DiffCamera. If we're going to be working on other stuff,
it's not needed.

Finally, I changed some get_ functions to be @property functions,
since I find that more syntactically pleasing.