Overhaul current load process
There were two issues with the previous load process.

~~~~~~~~~~~~~~~~~~~~~~~~~~~~

First, ALL of our load code was going to have to be crammed into the
load method of CityState. This wasn't necessarily bad but it was a
pain. We had to cram everything into one function and we had to keep
loading variables in the CityState. Again, not bad, but something
about it irked me.

So instead, I've implemented what I call a "load_chain", which is
sorta like a Factory object, if a Factory object was made up of
self-managing functions. As we add more steps to the load process, we
can add more functions to the load_chain. Hopefully that will help our
loading code stay maintainable.

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

The second issue with our load process was with, much to my
displeasure, the DiffCamera.

Not that it didn't work - it actually did exactly what it was
supposed to. The issue was that it didn't make much sense to do what
we were doing in the first place.

In this project, most of the tiles will not be changing from moment to
moment. The only advantage that the DiffCamera presented was a faster
load time, since we didn't load any sprites. In return, we took a
pretty large performance hit. Considering the trade-offs, I thought it
would be more effective to just render the world tiles as part of the
load process.