Surrender unto the Pyglet Resource Loader
Pyglet's resource loader suggests that code and assets should be
bundled together. I thought it would be much cleaner simply to have a
/assets directory where everything went.

This didn't work out to well, and resulted in me simply leaving random
copies of textures around wherever the code saw fit to look for them.
The end result was far messier than any of the alternatives.

After learning how to use the Pyglet Resource Loader, I've refactored
the project to group the assets with the relevant code.

As part of that, I moved most of the World code into it's own package
under the city package - world. Since most of this prototype so far
has focused on developing the world, most of the code has been
squirreled away into that package.

In addition, I've added two new files under city.world.assets:
terrain_detail.py and terrain_primary.py. Each one is associated with
a png image of the same name. The idea is that we'll group together
the image and the code for handling/describing that image (as the
pyglet resource loader originally suggested I should do).

Currently, both files provide Enums for selecting particular images
from the textures using Enum constants. This is much better than
selecting images from the texture using magic numbers, as we were
doing previously.

In the future, these files may also become responsible for loading
their associated images. It's my hope that, with this system, we can
provide a hook for future modders to add in custom/expanded tilesets.