mplement initial tile set management
Currently, we store our sprites/tiles on sheets. We manually assign
each tile an enumerated constant and track its coordinate in the
sheet grid.

But what happens if we want to expand the tile set - say, for detail
terrain? We could produce a new image but then we would have to
reassign EACH OF THE COORDINATES for EVERY SINGLE UPDATE. My internal
screaming recommends that we don't. We also needed a way to match the
raw tile types (stored as bytes) with their texture grid coordinates
(stored as tuples).

The solution is in the new TileSet class, which is meant to contain
one "set" or "class" of tiles. Currently, we have two sets: one for
"primary terrain" tiles (32 x 32) and one for "detail terrain" tiles
(16 x 16).

The greatest feature of the TileSet is that it allows us to combine
multiple ImageGrids (and their companion enums) into a single
semi-obfuscated class. It also manages assigning the companion enums
to a integer number designation so that all the separate enums can be
used without conflict.