Delete terrain painter
Originally, terrain_painter.py was supposed to be the "world painter"
that made decisions about what tiles should be which; maker.py would
be the coordinator that schedules the the painting methods.

It would be a separation of concerns - scheduling versus actual logic.

However, I realized that the coordinating logic consisted of about
80-some-odd lines of code. At that point, it was more like separation
for the purpose of having more files.

So now, terrain_painter.py has been removed and it's code has been
placed in maker.py.