Add detail map and edge generation
Added an extra world layer to keep track of - the detail map. This
layer will hold 16 x 16 details to add some extra visual detail to
the world.

To help handle this additional map layer (and any additional layers
yet to come) we've added a WorldData class. This mostly just serves
as storage for our raw and shaped world data. In the future, we may
have the WorldData class manage the sprite lists too.

As we now have 16 x 16 tiles designed to be used with 32 x 32 tiles,
the use16 functionality has been removed.

We also added our first use for the detail layer: edge passing, which
passes over the map and adds transition edges, which look quite
nice! Unfortuanetly, the process is pretty intensive in terms of
memory, so we'll need to fine tune the algorithm some more.