Switch to Tile classes instead of Sprites
Previously, we were using Pyglet's default Sprite class.

This was working well BUT we needed the ability to scale our sprites
up with repeating textures - which Sprite didn't readily offer.

Now, we've switched to using the custom-made Tile classes, which offer
less general functionality than the Sprite class but offer the more
specific functionality that we require.

To support this switch, CityState is playing host to an enum-keyed
dict containing all the texture groups. This is a prototype-only
band aid solution; handling the texture groups should be the
responsibility of whoever will be actively switching the textures.
Unfortunately, we don't know exactly where that responsibility will
land at this juncture.