Add tile averagingdd tile averaging
As a measure to cut down on load times and memory usage, I added a
system that averages together all the TERRAIN tiles in Z x Z areas
(currently 10 x 10). Instead of rendering each terrain tile in the
area, it renders one giant tile - the mode for that section - and then
renders non-mode tiles on top of that.

The result is pretty good - especially since current generation mostly
produces a grass land. We'll have to observe that as time goes on.
There is some mild frame skipping, so we'll have to look into that.

To support these changes, WorldData recieved a light rework. The World
Maker methods were also retooled - primarily to change their inputs to
WorldData objects (honestly, don't know why I did it any other way).