Add in FPS & time-diff display
I wasn't quite sure whether the hitching I was seeing was real or not,
so I've added in two separate GUI elements to help us track the
framerate and gamecycles.

It does seem like there is hitching happening occassionally, but it's
mostly minor. Maybe it's a memory issue? We'll see.