Add lake generation
Added in lake generation. The current method creates these stylized
sort of blobs composed of squares. This works pretty well, so we'll
leave it for now.

The reason we've generated lakes as these purpose-made blobs, rather
than a more realistic looking perlin noise shape, is so that we know
the location of each and every lake. This will be useful for the next
step, when we generate rivers. This method will allow us to use these
lakes as sources, and destinations, for generating rivers.