Change world data layers interface
Previously, the preferred method for changing the layer values in
world data was by getting a numpy array using some provided
get_Y_shaped methods.

This was working fine BUT it was problematic for "sympathetic"
changes. For example, whenever we changed the base terrain layer, the
average terrain layer also needed to be updated. This lead to three
separate lines that needed to be copy-pasted whenever we changed one
thing.

Now, we've introduced the WorldLayer classes to act as buffers for the
multiprocessing Arrays. This has enabled us to add extra feautres,
like sympathetic changes between different layers.

However, our performance has seen an impact. The "Base Terrain"
generation step now takes around 12 seconds instead of the usual... 1?
Can't remember. The "Edge Pass" now takes almost 45 seconds, up from
the original 15.

So we'll have to see how to correct this - we may end up using the
numpy array in speed critical instances, or improving the function of
WorldLayer, or simply reverting to the old system.