Add Fill Lake generation
I wanted to try something that looked a bit different from the
stylized square-lakes that I was previously generating.

Instead of generating lakes as a chain of sub-lakes, we would try
generating lakes as a big flood fill of a Perlin Noise map.

The result? It depends on where we land - but the most typical lakes
are long, stringy, and spread everywhere. In other words, we are
generating RIVERS instead of lakes.

Not what we wanted. BUT, I'll be keeping the code in. I'm sure that if
I were to change the Perlin variables or switch to a different way of
generating noise - maybe using sine or cosine waves? - that we could
get a better result. As it stands, I will stick with the LakeChain
method.