Revise for ChainLake generation for clarity
Previously, a ChainLake was created as a series of tuples vaguely
glued together in a Python list. It could best be described as a
"I Can't Believe it's Not a Class!" Python amalgamation.

That was functional, but it lacked clarity. With the addition of the
FillLake object, it seemed appropriate to update the chain lake
functions with their own object - the ChainLake object.