Build in program/theory tests
As time has gone on, I've been adding and removing different test
scripts. These aren't anything equivalent to formal tests; rather,
they typically test a theoretical functionality or experiment with an
existing component.

I wanted to include them in the project, but I also couldn't get the
tests to run anywhere except by just placing them in the src
directory. I didn't like that and was too apathetic to fix it, so the
tests went unadded.

Until now! The tests have been included in their own module, and are
meant to be run using the provided test.py script. Using the awesome
argparse library, the script will provide instructions on it's own
usage.