Add river generation phase
This update adds a river painting phase to the world making process.

Eventually, each lake that we generate will have a river coming from
it. For now, we are only generating a river off the first lake.

Generators are a new class of objects that encapsulate an (x, y) to
(value) algorithm. They were implemented specifically for rivers, but
they could have applications beyond the rivers.

Right now, rivers are generated by layering several Generators on top
of each other to produce a watershed heightmap, which the river then
follows.

The result is... not spectacular, but I'm sure with some tweaking it
can be improved.