Prototype a more directed river generator
After experimenting with the watershed test, I saw that the multiple
generators were more or less canceling each other out. Putting a slope
over a perlin noise map was mostly just turning everything to mush.

The idea was to have a directed flow, from a source, that still
followed a natural looking (i.e. windy) path. Perlin noise has the
perfect shape, but is less good for having a pre-defined start and end
point.

After seeing a post on GameDev Stack Exchange that suggested some
combination of perlin noise and A*, I was immediately inspired to
create a far worse solution.

Now, we have a incredibly dumb path finding turtle thing.

The turtle navigates a stable grid of points that map to a
randomly-shifted point on the real map. This grid allows us to more
easily detect river collisions (i.e. confluences) and provides a more
varied output (due to the randomly shifted points).

Once we have generated a river for each source, we take the results
(which are stored as vector-like objects called LatticePoints) and
create RiverSegments.

The result is janky and broken. The river appears weird and choppy,
broken and distorted. BUT - the rivers follow a meandering path to the
maps edge.

What I'm saying is, it's a start. With further refinement, I'm sure we
can beat this algorithm into a decent shape. Maybe even an elegant
one.