Increase legibility of river generation
In an attempt to make the river generation code easier to understand
and maintain, I've added a RiverLattice class that handles several of
the functions that made the river generation algorithm so fat and
massive.

This allowed for further tuning, and our rivers are now no longer
broken messes. They are now contigious, if a bit mismatched at the
junctions.

However, they are currently just a little bit too nice; they tend
towards long straight aways with little variation. The algorithm will need to be tweaked.