Convert LatticePoints to Segments via class method
Previously, we were converting LatticePoints to RiverSegments using a
class method of the RiverLattice.

This was fine, but I figured it would be more intuitive if the
LatticePoint object could convert itself to RiverSegments, rather than
having an external class (RiverLattice) do so.

I also took the time to change the algorithm around to be a bit more
neutral, and to use the nifty fractions library.

The result?

I made it worse. I made it so much worse. The algorithm is easier to
understand but the output is formed from the raw nightmare matter of
Picasso. I think I need some debugging.