Introduce Bresenham functionality
So, while searching for a solution to the funky rivers, I stumbled
over Bresenham functions - the line drawing algorithm and it's
extension, the midpoint circle algorithm.

Most of the examples presented online called paint on every (x, y)
pairing. To extend this functionality, I redid the algorithm to
accept a function with prepackaged args and keyword args. This turns
the Bresenham functions into a tool for applying functions to an area.

Since I figured these functions will have applications outside of just
rivers, I've included the functions under the game_util tab.