Experiment with Enum flexibility
So I like the usage of Enums for specifying tile types. However, we
need more information about each type of tile entry. Is it an
animation? What's it's precedence value? Does it have a proxy that
should be referred to instead of the enum itself?

To test this upcoming functionality, I've added a test that messes
around with creating several different varieties of Enum.

The results are pretty promising - we should be able to implement some
of my findings real soon.