Show visual flow direction for rivers
To assist with debugging, and lay the groundwork for a future river
flow animation, I've added in several river flow directional tiles.

Originally, I was basing the river's flow on the RiverFlow enum of the
segment. This led to some issues - the RiverFlow was based on the
river lattice, rather than the real positions of each lattice point.
This meant we frequently had rivers flowing downwards that were
actually flowing upwards.

To compensate, I quickly added a slapdash way of determining a
lattice point's flow direction based on the real positions of the two
lattice points.