Use unit circle methodology for river flow
In an effort to improve the river flow decision algorithm, I've
switched to using a method that compares the unit circle against the
slope between two positions to determine direction.

As part of this effort, I merged the river_dir_on_angle into the
generate_rivers method - I believe this is more appropriate than a
separate method. This necessitated a change in LatticePoint, since we
now wanted to track the LatticePoint flow as well as the real/visual
flow.

The result? It's a definite improvement. Right now our judgement is
geared ENTIRELY towards X, which is resulting in some weird patterns
when the change is occurring almost entirely on Y.