Decide river flow segment-by-segment
In previous iterations, I was attempting to assign the flow of the
river in vast swathes by assigning the flow to the LatticePoint
object.

However, our rivers do not consistently stay in the same orientation
for a LatticePoint. Sometimes the next segment of the river is to the
East, sometimes the South East - it changes from segment to segment.

Recognizing that fact, I have changed the algorithm to decide the flow
on a per-segment basis. The river flow decision now occurs in
LatticePoint's _make_segment method.