Create CardinalEnum from RiverFlow enum
While I was looking over the edge algorithm, I realized that one of
the elements that makes it so complicated and ugly is the use of a
"shifts" list.

I had (inadvertently) already come up with a better solution with the
RiverFlow enum. Realizing the potential use for that enum, I moved it
to the game_util package and have transformed it into the
CardinalEnum. The CardinalEnum has also consumed the ShiftEnum, since
I couldn't really justify it being separate.