Revise WorldData tile interface and averaging
So I thought that just accessing the multiprocessing array directly
would be indistinguishable from accessing it through numpy. I also
thought it would be an improvement to use a ctypes Structures for our
averaging.

The trade off with how we were doing averaging: the processing/paint
phase went slower, but the render step went quicker.

However, after experimenting a little, I reached a couple of
conclusions. First, using the multiprocessing array directly - rather
than with numpy - didn't seem to have any improvement in processing
time. In fact, it seemed to make it worse (though I didn't verify
that).

To correct this, we now use numpy in the WorldLayer data class.

Secondly, using the ctypes Structures for storing our averages was
making everything VERIFIABLY slower. It seems that accessing the
fields of a ctypes struct isn't that efficient.

As I was changing back to the old implementation of averaging, I
realized that the "count types as you go" methodology wasn't actually
saving us any time. Just counting all the averages in one pass seems
to be SAVING us time!

So we now have an "average" layer where the only information we store
is the average tile for each section - no running counts. We now
determine the average in one pass during the world maker step.