Add the framework for an edge test
We're going to try doing some work on the edges; in order to test
that, we're going to need to have an edge test.

In order to support this test, we needed to modify some of our
existing classes. Primarily, the load_chain and CityState needed to be
modified so that we could pass in a world_maker function of our
choice.

To test this, we've added in the edge_test.py file - and it seems to
work fine so far. That being said, the current systems seem to be
rebelling against the small map size, so we might need to fix that.