Complete the edge test
After we layed down the groundwork for testing the edges, I've built
up the edge test.

This edge test, and it's two painting methods, will provide a good
basis for evaluating our edge algorithms in the future.

While I created this edge test specifically so I could test adding
additional edges, I discovered that having so many edges at once was
a massive framerate bomb. With that knowledge in mind, I think I'm
going to leave the edge painting as it stands right now, and move on.

It could be that the particular density of edges was much higher than
it would be in the final product, and that this test is thusly not
indicative of what the final game would be like. But I think it will
suffice for me.