Create a framework for distributed random tiles
So we want detail tiles to be placed randomly. Except not exactly
randomly.

We want contigious (i.e. Perlin) noise to help determine where the
tiles should be placed - so that you get parts with lots of detail and
parts with little detail.

However, we don't just want purely random tile placement - we want
there to be a sort of contigious sense of order to what tiles get
placed as well.

Our solution is to use Voroni polygons, where each polygon specifies a
tile distribution. Thus we get continuity for both the location and
the tile typing.

To implement this, I've added a VoroniDistro class. It takes in a
list-tuple thing and magically transmutes it into two lists using
Pythonic black magic channeled straight from Guido himself. In other
words, I used a zip function.

I then created several sample VoroniDistros for us to use. We'll
implement the actual Voroni algorithm next.