Add in detail tiles
Added in some detail tiles, in order to add some visual distinction to
the vast fields of green that currently dominate the landscape. This
adds just a hint of variation. We'll propbably reuse this algorithm
when we go about placing extra details like trees.

However, the framerate has taken another dip, and the loadtime has
taken another hike. I begin to doubt the fesability of this venture...