Move the turret using PHYSICS (or a derivative...)
I had a very specific idea for how I wanted the turret to move. This
update adds acceleration and decceleration in order to better approach
that ideal.

I'm actualy very satisfied with the FIZZIKS equation I devised - the
effect it has, not the god awful formula itself. One thing I noticed
is that, while you can pretty much get the turret aiming correctly,
absolute percision is not achievable (with things the way they are).

We'll have to keep that in mind - especially when it comes to enemies.