Begin experimenting with Python alternatives
Python was one of the pillars this project was founded on. But
honestly, after working with it for so long, I just don't know if it
has the speed to perform how I need it to. I downloaded a very basic
asteroids clone written in pyglet and it was CHUGGING the entire time.

Maybe it is possible. I just can't see it being worth the effort,
given the alternatives out there. But what are the alternatives?

I started by investigating the great titan, Unity. My only major gripe
was that Unix development support - my primary development platform -
is not... great. You can do it. Kind of.

While searching how to get Unity on to my Linux laptop, one comment
suggested using Godot Engine instead. It's very Unity-esque, deploys
to multiple platforms, and can be developed on Linux!

And, as the icing on the cake, it's built-in scripting language is
Python-inspired. So, after following some tutorials, I'm ready to try
something very basic. Two character in a room, together. Let's see
what I can do.