Add in a turret with some basic controls (Godot)
To catch up with our Python prototype, I've added in a turret that
rotates with the Q and E keys. This puts us on par with the
development of our Prototype II Python development.

That I was able to duplicate my Python development so quickly is a
testament to Godot. However, I am struggling with the engine's system
of nodes.

In Pyglet, and cocos2d, a Sprite would be responsible for detecting
collisions, physics and the like. Here, that isn't so - a Sprite is
only for rendering. To add collisions, you need an Area2D node. To
check it's visibility, you need a VisibilityNotifier node. To add
physics, you need a KinematicBody node, and so forth.

This more mix-and-match instances wouldn't be too bad but I'm my usual
techniques (googling for examples and guides) just doesn't work, which
is slowing my progress. Maybe I should start picking apart some
templates...