Fix the GUI scaling issue in Main
Previously, I was having a massive problem with the GUI in the Main
scene. Even though I had set it up very precisely, the "layout"
options weren't doing anything and it wasn't scaling when stretched
and squashed.

Turns out the problem was actually with the root node of our main
scene. It was a Node2D instead of.. just a regular Node. Not sure I
entirely understand why there's a dramatic difference between those
two, but that was the difference between a broken scene and a working
one.

I suspect it's because Node2D is supposed to be for moving elements -
maybe? Like Sprites and Rigidbody's and stuff? That could be why the
layout button was orienting everything around the origin, instead of
where it was supposed to be.