Add some basic information to the Wheel display
So I eventually want to rotate the wheel as our input for rotating the
turret. Or maybe we'll just use it to show how the turret is rotating.
Not sure yet.

Either way, I've added some extra symbols and wording to indicate
directions on the wheel - with abbreviations for clockwise and
counter-clockwise.