Begin developing a turnable wheel
So I want the user to interact with the wheel by clicking and dragging
the wheel circularly. Step one of that plan is being able to detect
whether the wheel has been clicked.

To facilitate that, the wheel was changed to a texture button.
A thematic label - which indicates whether the button is pressed or
not - has also been added. The GUI/GUIBox node now emits signals for
interfacing with the wheel when it's pressed up or down.