Refactor (scuttle) our signal set up
So I was thinking that we would handle rotation for the wheel at the
top-level, in the game node. But then I realized that this was
antiquated thinking on my part.

Even though the turret needs to respond to the same inputs - or rather
respond in kind to input from a different node. However, I realized
that this was breaking the design philosophy of Godot.

Rather, I should handle input in the GUIBox, then emit a signal to
indicate that the wheel turned, and to what degree. That's the Godot
way!

This allowed me to renovate (delete) some existing scripts.