Connect the wheel's rotation to the turret
I connected the rotation of the GUI's traverse wheel to the rotation
of the turret.

The result is cool, but kind of clunky from a control perspective. I
don't think I'll use this idea after this prototype but learning how
to do it was definitely a good introduction to Godot.

One problem I've been having has involved scripts, signals, and
scenes. From what I can tell, an instanced scene obscures all it's
children nodes (save the root) from the scene that's instanced it.
That's not my issue - that's just good object-oriented programming.

My issue is that we have children nodes of the instanced scene that
are emitting signals. For neatness, I keep the emitting in the scene
itself - but then how do we make signals of children nodes accessible
to nodes outside the scene?

I tried a "pass-through" scheme, where the root node's script listened
for signals and then just emitted a basically duplicate signal. This
was going to lead to a bunch of unecessary functions, so I skipped
that.

My current solution is to offer "Registration Methods" that allow us
to manually register whatever signals we want with the children nodes.
This works, but I PREFER doing things through the GUI - it feels more
official and clear. I'll take what I can get, though.