Add a shell path polygon to our turret
Given the sort of scale that I'm aiming for towards the end of this
prototype, the turret itself is going to be very tiny. Using the
barrel as the primary aiming device would be... ineffective.

To compensate, I've added a 'ShellPath', which illustrates the path
the shell will take. I've also added controls for increasing and
decreasing the path's length.

I also moved the Turret scene into it's own folder - so despite what
the commit claims, I didn't REALLY delete it.