Add a button workaround for the Traverse Wheel
While playing around, I found that pressing the Spacebar after
clicking the Traverse wheel allowed you to engage the lay process
without having to click the button. This seemed to be because of the
'ui_select' action. I really liked that functionality.

Godot has a functionality that allows us to duplicate this -
Shortcuts. You can simply hook up an input action and it will press
the button when the action is pressed.

This would be great if it wasn't demonstrably broken. Searching
around, it seems like this was broken in version 3.0 and was supposed
to be fixed in version 3.1, where it is still broken.

Therefore, we have no choice but to check our two conditions - the
button press and the action press - in our _process function. It's
inefficient and that BUGS me but without the Shortcuts I don't have
much of a choice.