Add "Range" GUI section
Added a section to the GUI concerning increasing and decreasing the
range. I haven't bothered creating any code for this particular
section - I think I'm going to fill up the rest of the GUI first.

This section will be easier than the last one - we'll bind the two
buttons to Q and E and use the progress bar to represent the current
range.