Switch TGLP label from hide/show to a color change
Previously, we were hiding and showing the TGLP label to indicate the
current status. This was fine, but I felt it wasn't thematic enough.

It would be better, I think, to simply have the label fade somewhat,
so that it appears as an indicator that is simply unlit. I like it
better this way.