Add a "Control" section to the GUI
Added a fire control section to the GUI. This includes a fire
button (which fires the turret), a status label that we'll change,
and a gun bolt.

Keeping with the theme of a kinetic / interactive GUI, the player will
have to drag the bolt left to unload, then drag the bolt right to
reload the gun. I want to grab a nice clink-clink sound for whenever
we reload.