Added the final GUI section - "Payload"
Added a final section to the GUI for controlling the turret's payload.
This section will allow the player to change between two different
payload types - Phosphorous and Explosive.

I used an animated sprite for the payload's icon. I feel like that's
not the best way to do a texture swap on Rects of the same size, but
whatever