Connect the Range Bar controls to the turret
Connected the range controls to the turret. I'm definitely beginning
to learn more and more about how a Godot project is structured.

I'm not quite sure how I would do it better, but I know the way I'm
doing it right now isn't optimal. Right now, we have a Main node that
has a Turret node and a GUI node. Some values are stored in the GUI,
others in the Main, and others in the Turret. We then have to shift
and sync the values between this various nodes.

For a small prototype, that probably won't cause too many issues but
I can see how in a larger project - including this project down the
line - that we could end up with an absolute bramble nightmare.

We need to keep track of how data is flowing - because the flow down,
right, left, up, and back again could cause us some real pain.