Fully switch over to a Godot project
I've finally committed to using Godot for our project. I've deleted
the Python source files, but I did preserve the texture images I had
made.

My plan for this prototype is to try and create an arena brawler.
Mind you, the player is not in control of the arena brawler, merely
an observer.