Experiment with using tiles from an AutoTile set
So I experimented with how exactly we go about placing down tiles from
an Atlas/Autotile.

Part one of that was changing PrimaryTileSet to using single tiles to
identify the tiles. That itself wasn't too bad.

I then tried doing that with DetailTileSet, which made me go slightly
insane. So I had to do it with an AutoTile/Tileset; but how would I do
that?

After some searching, I found that you can provide the coordinates
of the AutoTile tile to the Map in the set_cell() method.
Unfortunately, while there are optional arguments in Godot, there is
no named arguments, so in order to provide that argument you have to
give a whole bunch of unnecessary arguments. Oof!

Okay, so here's my idea: we'll have a singleton, WorldTiles.gd, that
contains the constants for accessing the tiles. It will be pretty
similar to the asset describing scripts from when we were using
Python, complete with enumerations of the keys. I'll also have to
devise some sort of way of easily accessing detail tiles and corner
tiles more easily, based on the counterpart primary tile.

My new plan for our WorldBase class (renamed from World) is that it's
script will contain the basic/common functions that any subsequent
specialized worlds would require. Right now, I'm thinking that will
include performing the edge determination, and also the selection of
detail tiles given a location.

Also, just a quick note: the detail tiles were getting some of the
tile above them so they are now offset by -1. The Godot working view
of the map shows artifacting and other issues, but these are not
present during actual testing.