Add world scrolling scene
Added a world scrolling scene. In the forth-coming steps we'll be
starting our world generation. As we stress test Godot's handling for
generation, we'll be stress testing the world size and will need to be
able to scroll over the world. This little scene should facilitate
that.

One thing I've noticed - there's a lot of input lag when scrolling the
camera. I wonder why that is? I'll ignore it for now, but we might
need to look into what's happening there later on...